*: move "lease" to "internal/lease"
