+++ etcdserver/api/v3rpc/lease.go
@@ -21,7 +21,7 @@ import (
 	"github.com/coreos/etcd/etcdserver"
 	"github.com/coreos/etcd/etcdserver/api/v3rpc/rpctypes"
 	pb "github.com/coreos/etcd/etcdserver/etcdserverpb"
-	"github.com/coreos/etcd/lease"
+	"github.com/coreos/etcd/internal/lease"
 )
 
 type LeaseServer struct {
