+++ etcdserver/backend.go
@@ -19,8 +19,8 @@ import (
 	"os"
 	"time"
 
+	"github.com/coreos/etcd/internal/lease"
 	"github.com/coreos/etcd/internal/raftsnap"
-	"github.com/coreos/etcd/lease"
 	"github.com/coreos/etcd/mvcc"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/raft/raftpb"
