+++ etcdserver/server.go
@@ -38,10 +38,10 @@ import (
 	"github.com/coreos/etcd/internal/alarm"
 	"github.com/coreos/etcd/internal/compactor"
 	"github.com/coreos/etcd/internal/discovery"
+	"github.com/coreos/etcd/internal/lease"
+	"github.com/coreos/etcd/internal/lease/leasehttp"
 	"github.com/coreos/etcd/internal/raftsnap"
 	"github.com/coreos/etcd/internal/store"
-	"github.com/coreos/etcd/lease"
-	"github.com/coreos/etcd/lease/leasehttp"
 	"github.com/coreos/etcd/mvcc"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/pkg/fileutil"
