+++ etcdserver/v3_server.go
@@ -23,8 +23,8 @@ import (
 	"github.com/coreos/etcd/auth"
 	pb "github.com/coreos/etcd/etcdserver/etcdserverpb"
 	"github.com/coreos/etcd/etcdserver/membership"
-	"github.com/coreos/etcd/lease"
-	"github.com/coreos/etcd/lease/leasehttp"
+	"github.com/coreos/etcd/internal/lease"
+	"github.com/coreos/etcd/internal/lease/leasehttp"
 	"github.com/coreos/etcd/mvcc"
 	"github.com/coreos/etcd/raft"
