+++ internal/lease/lessor.go
@@ -22,7 +22,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/coreos/etcd/lease/leasepb"
+	"github.com/coreos/etcd/internal/lease/leasepb"
 	"github.com/coreos/etcd/mvcc/backend"
 )
