+++ mvcc/kv_view.go
@@ -14,9 +14,7 @@
 
 package mvcc
 
-import (
-	"github.com/coreos/etcd/lease"
-)
+import "github.com/coreos/etcd/internal/lease"
 
 type readView struct{ kv KV }
