+++ tools/benchmark/cmd/mvcc-put.go
@@ -21,7 +21,7 @@ import (
 	"runtime/pprof"
 	"time"
 
-	"github.com/coreos/etcd/lease"
+	"github.com/coreos/etcd/internal/lease"
 	"github.com/coreos/etcd/pkg/report"
 
 	"github.com/spf13/cobra"
