+++ tools/etcd-dump-db/backend.go
@@ -20,7 +20,7 @@ import (
 	"path/filepath"
 
 	bolt "github.com/coreos/bbolt"
-	"github.com/coreos/etcd/lease/leasepb"
+	"github.com/coreos/etcd/internal/lease/leasepb"
 	"github.com/coreos/etcd/mvcc"
 	"github.com/coreos/etcd/mvcc/backend"
 	"github.com/coreos/etcd/mvcc/mvccpb"
